Initial commit
chore: Add caching for user data in UsersController

The UsersController has been modified to include caching for user data using the `@nestjs/cache-manager` module. The cache is set to expire after 24 hours. This change improves performance by reducing the number of requests made to the microservice.

This commit also updates the UsersModule to import the RedisModule for caching.

Note: The recent repository commits do not provide any relevant information for generating the commit message.

chore: Update cache TTL in UsersController

The UsersController has been updated to use the `envs.cacheTtl` value instead of `envs.cacheTime` for setting the cache TTL. This change ensures that the cache expires after the specified time (24 hours) for improved performance and data freshness.